Refactor naming and move around Classes into folders
Move Microsoft To Do related files into its own folder so that it can be
 easily distinguished from other platforms including Google Tasks
