+++ CLI To Do/GoogleTasks/GoogleTaskHelper.cs
@@ -1,0 +1,5 @@
+﻿namespace CLI_To_Do.GoogleTasks; 
+
+public class GoogleTaskHelper {
+    
+}
